Updated generator

+++ generate.sh
@@ -1,20 +1,23 @@
-# Note: This can take a while to run (about a minute for me)
+if [ -d "vrchat_dart_generated" ]; then
+    echo "Removing old files"
+    cd vrchat_dart_generated
+    # Remove old files
+    rm -rf doc lib test
+    cd ..
+fi
+
+cd vrchat_dart
+dart run tool/patch_input.dart
+cd ..
+
+openapi-generator generate -g dart-dio -c generator-config.yaml
+
+# cd vrchat_dart
+# dart run tool/patch_output.dart
+# cd ..
 
 cd vrchat_dart_generated
-# Remove old files
-rm -rf doc lib test
-
-cd ../vrchat_dart
-
-# For some reason the openapi generator won't run without a clean
-flutter clean
-flutter pub get
-
-flutter pub run build_runner build --delete-conflicting-outputs
-
-cd ../vrchat_dart_generated
+dart pub get
+dart run build_runner build --delete-conflicting-outputs
 dart fix --apply
-flutter format .
-
-cd ..
-dart run vrchat_dart/tool/patch.dart
+dart format .
